useropts.wdrestore
When a task is run, the wd is set to the project root. If the task runs
in a terminal, the original buffer will lose its working directory,
because there's no time to restore it, since the focus goes immediately
to the terminal. So I set up an autocommand to restore it when
re-entering the original buffer.